requirements.txt
fastapi~=0.118.0
uvicorn
sqlalchemy~=2.0.43
psycopg2-binary
alembic~=1.16.5
python-jose~=3.5.0
passlib[bcrypt]~=1.7.4
pydantic~=2.11.9
redis~=6.4.0
pytest~=8.4.2
pydantic-settings~=2.11.0
email-validator
python-multipart
python-dotenv~=1.1.1
pip~=25.2
typing_extensions~=4.15.0
packaging~=25.0